Migrations, and caught some edge cases

+++ app/app.py
@@ -110,15 +110,27 @@ def create_app(config_name):
 
 	def middleware():
 		auth_header = request.headers.get('Authorization')
-		access_token = auth_header.split(" ")[1]
-		if access_token:
-			user_id = Users.decode_token(access_token)
-			return user_id
+		if auth_header:
+			access_token = auth_header.split(" ")[1]
+			if access_token:
+				user_id = Users.decode_token(access_token)
+				if not isinstance(user_id, int):
+					response = {
+						'status': 'fail',
+						'message': user_id
+					}
+					return response, 403
+				return user_id
+			response = {
+		    	'status': 'fail',
+		    	'message': 'Bearer token malformed.'
+		    }
+			return response, 401
 		response = {
-	    	'status': 'fail',
-	    	'message': 'Bearer token malformed.'
-	    }
-		return response, 401
+			'status': 'fail',
+			'message': 'Authorization is not provided'
+		}
+		return response, 500
 
 	def is_valid(value, min_length, _type):
 		message = []
@@ -137,9 +149,8 @@ def create_app(config_name):
 	class ShoppingListAPI(Resource):
 
 		def get(self):
-			user_id = middleware()
-			
-			if not isinstance(user_id, str):
+			user_id = middleware()					
+			if isinstance(user_id, int):				
 				shoppinglists = ShoppingList.query.filter_by(owner_id=user_id)
 				results = []
 				for shoppinglist in shoppinglists:
@@ -152,16 +163,17 @@ def create_app(config_name):
                         'owner_id': shoppinglist.owner_id
                     }
 					results.append(obj)
-
+				if len(results) == 0:
+					response = {
+						'message': "You don't have any shoppinglists for now"
+					}
+					return response, 200
 				response = jsonify(results)
 				response.status_code = 202
 				return response
 			else:
                 # Return token error message
-				response = {
-                    'message': user_id
-                }
-				return response, 401
+				return user_id
 		def post(self):
 			post_data = ['title', 'description']
 			for arg in range(len(post_data)):
@@ -220,20 +232,22 @@ def create_app(config_name):
 			return response, 202
 
 		def put(self, shoppinglist_id):
-			post_data = ['title', 'description']
-			for arg in range(len(post_data)):
-				parser.add_argument(post_data[arg])
-			args = parser.parse_args()
-			title = args['title']
-			description = args['description']
-			valid_title = is_valid(value=title, min_length=10, _type="text")
-			valid_description = is_valid(value=description, min_length=10, _type="text")
-			if valid_title is True and valid_description:
-				user_id = middleware()
-				if not isinstance(user_id, str):
-					check_exists = ShoppingList.query.filter_by(title=title).first()
-					if check_exists is None:
-						shoppinglist = ShoppingList(title=title, description=description, owner_id=user_id)
+			user_id = middleware()
+			if isinstance(user_id, int):
+				post_data = ['title', 'description']
+				for arg in range(len(post_data)):
+					parser.add_argument(post_data[arg])
+				args = parser.parse_args()
+				title = args['title']
+				description = args['description']
+				valid_title = is_valid(value=title, min_length=10, _type="text")
+				valid_description = is_valid(value=description, min_length=10, _type="text")			
+				if valid_title is True and valid_description is True:
+					title_exists = ShoppingList.query.filter_by(title=title).first()
+					if not title_exists:
+						shoppinglist = ShoppingList.query.filter_by(owner_id=user_id, id=shoppinglist_id).first()
+						shoppinglist.title = title
+						shoppinglist.description = description
 						shoppinglist.save_shoppinglist()
 						# Return Response
 						response = {
@@ -247,15 +261,22 @@ def create_app(config_name):
 					response = {
 						'message': 'Shopping List {} already exists'.format(title)
 					}
-					return response, 202
-				response = jsonify({
-					'message': user_id
-				})
-				return response
-			response = {
-				'message': valid_title
-			}
-			return response, 202
+					return response, 202					
+				else:
+					if valid_title is True:
+						message = valid_description
+					elif valid_description is True:
+						message = valid_title
+					else:
+						message = [valid_title, valid_description]
+					response = jsonify({
+						'message': message,
+						'status_code': 202
+					})
+					return response
+			else:
+                # Return token error message
+				return user_id
 
 		def delete(self, shoppinglist_id):
 			shoppinglist = ShoppingList.query.filter_by(id=shoppinglist_id).first()
@@ -273,10 +294,9 @@ def create_app(config_name):
 	class ShoppingListItemsAPI(Resource):
 
 		def get(self, shoppinglist_id):
-			user_id = middleware()
-			
-			if not isinstance(user_id, str):
-				shoppinglistitems = ShoppingListItem.query.filter_by(shoppinglist_id=shoppinglist_id)
+			user_id = middleware()			
+			if isinstance(user_id, int):
+				shoppinglistitems = ShoppingListItem.query.filter_by(shoppinglist_id=shoppinglist_id, owner_id=user_id)
 				results = []
 				for shoppinglistitem in shoppinglistitems:
 					obj = {
@@ -289,27 +309,30 @@ def create_app(config_name):
                         'owner_id': shoppinglistitem.owner_id
                     }
 					results.append(obj)
+				if len(results) == 0:
+					response = {
+						'message': "You don't have any items for now"
+					}
+					return response, 200
 				response = jsonify(results)
-				response.status_code = 200
+				response.status_code = 202
 				return response
 			else:
                 # Return token error message
-				response = {
-                    'message': user_id
-                }
-				return response, 401
+				return user_id
+
 		def post(self, shoppinglist_id):
-			post_data = ['item_title', 'item_description']
-			for arg in range(len(post_data)):
-				parser.add_argument(post_data[arg])
-			args = parser.parse_args()
-			item_title = args['item_title']
-			item_description = args['item_description']
-			valid_item_title = is_valid(value=item_title, min_length=10, _type="text")
-			valid_item_description = is_valid(value=item_description, min_length=10, _type="text")
-			if valid_item_title is True and valid_item_description:
-				user_id = middleware()
-				if not isinstance(user_id, str):
+			user_id = middleware()
+			if user_id:
+				post_data = ['item_title', 'item_description']
+				for arg in range(len(post_data)):
+					parser.add_argument(post_data[arg])
+				args = parser.parse_args()
+				item_title = args['item_title']
+				item_description = args['item_description']
+				valid_item_title = is_valid(value=item_title, min_length=10, _type="text")
+				valid_item_description = is_valid(value=item_description, min_length=10, _type="text")
+				if valid_item_title is True and valid_item_description is True:
 					check_exists = ShoppingListItem.query.filter_by(item_title=item_title).first()
 					if check_exists is None:
 						shoppinglistitem = ShoppingListItem(item_title=item_title, item_description=item_description, shoppinglist_id=shoppinglist_id, owner_id=user_id)
@@ -328,52 +351,65 @@ def create_app(config_name):
 						'message': 'Shopping List item {} already exists'.format(item_title)
 					}
 					return response, 202
-				response = jsonify({
-					'message': user_id
-				})
-				return response
-			response = {
-				'message': valid_item_title
-			}
-			return response, 202
+				else:
+					if valid_item_title is True:
+						message = valid_item_description
+					elif valid_item_description is True:
+						message = valid_item_title
+					else:
+						message = [valid_item_title, valid_item_description]
+					response = jsonify({
+						'message': message,
+						'status_code': 202
+					})
+					return response
+			else:
+                # Return token error message
+				return user_id
 
 	class SingleShoppingListItemAPI(Resource):
-
 		def get(self, shoppinglist_id, shoppinglistitem_id):
-			shoppinglistitem = ShoppingListItem.query.filter_by(
-				item_id=shoppinglistitem_id,
-				shoppinglist_id=shoppinglist_id
-			).first()
-			if shoppinglistitem:
+			user_id = middleware()
+			if isinstance(user_id, int):
+				shoppinglistitem = ShoppingListItem.query.filter_by(
+					item_id=shoppinglistitem_id,
+					shoppinglist_id=shoppinglist_id
+				).first()
+				if shoppinglistitem:
+					response = {
+						'item_id': shoppinglistitem.item_id,
+						'owner_id': shoppinglistitem.owner_id,
+						'shoppinglist_id': shoppinglistitem.shoppinglist_id,
+						'item_title': shoppinglistitem.item_title,
+						'item_description': shoppinglistitem.item_description,
+						'message': 'success'
+					}
+					return response, 201
 				response = {
-					'item_id': shoppinglistitem.item_id,
-					'owner_id': shoppinglistitem.owner_id,
-					'shoppinglist_id': shoppinglistitem.shoppinglist_id,
-					'item_title': shoppinglistitem.item_title,
-					'item_description': shoppinglistitem.item_description,
-					'message': 'success'
+					'message': 'Requested value \'{}\' was not found'.format(shoppinglistitem_id)
 				}
-				return response, 201
-			response = {
-				'message': 'Requested value \'{}\' was not found'.format(shoppinglistitem_id)
-			}
-			return response, 202
+				return response, 202
+			else:
+                # Return token error message
+				return user_id
 
 		def put(self, shoppinglist_id, shoppinglistitem_id):
-			post_data = ['item_title', 'item_description']
-			for arg in range(len(post_data)):
-				parser.add_argument(post_data[arg])
-			args = parser.parse_args()
-			item_title = args['item_title']
-			item_description = args['item_description']
-			valid_item_title = is_valid(value=item_title, min_length=10, _type="text")
-			valid_item_description = is_valid(value=item_description, min_length=10, _type="text")
-			if valid_item_title is True and valid_item_description:
-				user_id = middleware()
-				if not isinstance(user_id, str):
+			user_id = middleware()
+			if isinstance(user_id, int):
+				post_data = ['item_title', 'item_description']
+				for arg in range(len(post_data)):
+					parser.add_argument(post_data[arg])
+				args = parser.parse_args()
+				item_title = args['item_title']
+				item_description = args['item_description']
+				valid_item_title = is_valid(value=item_title, min_length=10, _type="text")
+				valid_item_description = is_valid(value=item_description, min_length=10, _type="text")
+				if valid_item_title is True and valid_item_description is True:
 					check_exists = ShoppingListItem.query.filter_by(item_title=item_title).first()
 					if check_exists is None:
-						shoppinglistitem = ShoppingListItem(item_title=item_title, item_description=item_description, owner_id=user_id, shoppinglist_id=shoppinglist_id)
+						shoppinglistitem = ShoppingListItem.query.filter_by(owner_id=user_id, item_id=shoppinglistitem_id, shoppinglist_id=shoppinglist_id).first()
+						shoppinglistitem.item_title = item_title
+						shoppinglistitem.item_description = item_description
 						shoppinglistitem.save_shoppinglistitem()
 						# Return Response
 						response = {
@@ -386,28 +422,39 @@ def create_app(config_name):
 					response = {
 						'message': 'Shopping List {} already exists'.format(item_title)
 					}
-					return response, 202
-				response = jsonify({
-					'message': user_id
-				})
-				return response
-			response = {
-				'message': valid_item_title
-			}
-			return response, 202
+				else:
+					if valid_item_title is True:
+						message = valid_item_description
+					elif valid_item_description is True:
+						message = valid_item_title
+					else:
+						message = [valid_item_title, valid_item_description]
+					response = jsonify({
+						'message': message,
+						'status_code': 202
+					})
+					return response
+			else:
+                # Return token error message
+				return user_id
 
 		def delete(self, shoppinglist_id, shoppinglistitem_id):
-			shoppinglistitem = ShoppingListItem.query.filter_by(item_id=shoppinglistitem_id, shoppinglist_id=shoppinglist_id).first()
-			if shoppinglistitem:
-				shoppinglistitem.delete_shoppinglistitem()
+			user_id = middleware()
+			if isinstance(user_id, int):
+				shoppinglistitem = ShoppingListItem.query.filter_by(item_id=shoppinglistitem_id, owner_id=user_id, shoppinglist_id=shoppinglist_id).first()
+				if shoppinglistitem:
+					shoppinglistitem.delete_shoppinglistitem()
+					response = {
+						'message': 'Shopping list item \'{}\' deleted successfuly'.format(shoppinglistitem.item_title)
+					}
+					return response, 201
 				response = {
-					'message': 'Shopping list item \'{}\' deleted successfuly'.format(shoppinglistitem.item_title)
+					'message': 'Requested value \'{}\' was not found'.format(shoppinglistitem_id)
 				}
-				return response, 201
-			response = {
-				'message': 'Requested value \'{}\' was not found'.format(shoppinglistitem_id)
-			}
-			return response, 202
+				return response, 202
+			else:
+                # Return token error message
+				return user_id
 	
 	api.add_resource(Home, '/')
 	api.add_resource(Register, '/auth/register')
